taipan_launcher/requirements.txt
ttkbootstrap
pywebview
requests
